+++ packages/graphback-core/src/graphql/graphqlUtils.ts
@@ -29,24 +29,3 @@ export const filterObjectExtensions = (types: InputModelTypeContext[]) => types.
 export const createImplementsInterfaceString = (names: string[]) => {
   return `implements ${names.map((name: string) => name).join(' & ')} `;
 }
-
-/**
- * Collects all GraphQL files in a directory and reads the content into a string.
- *
- * @param schemaDir - The directory of the schema file(s)
- */
-export const buildSchemaText = (schemaDir: string): string => {
-  const schemaPath = join(schemaDir, '*.graphql');
-  const files = sync(schemaPath);
-
-  if (files.length === 0) {
-    return undefined;
-  }
-
-  const schemaText = files
-    // tslint:disable-next-line: no-unnecessary-callback-wrapper
-    .map((f: string) => readFileSync(f))
-    .join('\n');
-
-  return schemaText.length ? schemaText : undefined;
-}
